+++ api/BUILD
@@ -17,6 +17,5 @@ package(default_visibility = ["//visibility:public"])
 cc_library(
     name = "api",
     hdrs = glob(["include/**/*.h"]),
-    srcs = glob(["src/**/*.cc"]),
     strip_include_prefix = "include",
 )
